requirements.txt
hachoir
pillow
